changes based on pull request recommendations
