+++ package.json
@@ -38,9 +38,7 @@
     "prop-types": "^15.8.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-markdown": "^8.0.3",
-    "react-modal": "^3.16.1",
-    "remark-gfm": "^3.0.1"
+    "react-modal": "^3.16.1"
   },
   "devDependencies": {
     "@babel/preset-react": "^7.18.6",
