+++ pages/contact-us/contact-employment-insurance.js
@@ -3,7 +3,8 @@ import { Heading, TableContent } from '@dts-stn/service-canada-design-system'
 import { Fragment } from 'react'
 import en from '../../locales/en'
 import fr from '../../locales/fr'
-import ContactSection from '../../components/ContactSection'
+import ContactSection from '../../components/contact/ContactSection'
+import ContactProvince from '../../components/contact/ContactProvince'
 import { getProfileContent } from '../../graphql/mappers/profile'
 import { getBetaBannerContent } from '../../graphql/mappers/beta-banner-opt-out'
 import { getBetaPopupExitContent } from '../../graphql/mappers/beta-popup-exit'
@@ -12,7 +13,6 @@ import logger from '../../lib/logger'
 import Modal from 'react-modal'
 import React from 'react'
 import ExitBetaModal from '../../components/ExitBetaModal'
-import ContactProvince from '../../components/ContactProvince'
 import Markdown from 'markdown-to-jsx'
 const PageData = require('../../json/Nov30Data.json')
 
@@ -167,11 +167,6 @@ export default function ContactEmploymentInsurance(props) {
 }
 
 export async function getStaticProps({ res, locale }) {
-  const content = await getProfileContent().catch((error) => {
-    logger.error(error)
-    //res.statusCode = 500
-    throw error
-  })
   const bannerContent = await getBetaBannerContent().catch((error) => {
     logger.error(error)
     // res.statusCode = 500
